src/backend/app/handlers/geo_ip_handler.py
"""
Geo IP handler for CloudFlare headers
"""

from guard.protocols.geo_ip_protocol import GeoIPHandler


class CFHeaderGeoIPHandler(GeoIPHandler):
    """
    Custom GeoIPHandler class for handling geolocation based on IP addresses.
    """

    def __init__(self) -> None:
        self._initialized = False

    @property
    def is_initialized(self) -> bool:
        """
        Check if the geolocation handler has been initialized.

        :return: True if initialized, False otherwise.
        """
        return self._initialized

    async def initialize(self) -> None:
        """
        Initialize the geolocation handler.

        :return:
        """
        self._initialized = True

    async def initialize_redis(self, redis_handler: "RedisManager") -> None:
        """
        Initialize the Redis connection for geolocation.

        :param redis_handler: The Redis manager instance.
        :return:
        """
        return None

    def get_country(self, ip: str) -> str | None:
        """
        Extracts the country code from the given IP address.

        :param ip: The IP address to extract the country code from.
        :return: The country code associated with the IP address, or None if not found.
        """
        # TODO: Implement geolocation logic to get the country code
        # IDEA: Use a geolocation API or Redis database to retrieve the country code based on the IP address
